(Re)set globals on test setup, not on adapter load
One of these globals is overridden_error_class. It is important to delay setting
this because when we come to add the MiniTest / ActiveSupport adapter
(upcoming), it *and* the MiniTest adapter will actually have been loaded, and we
need to set different error classes in each case, and since the MiniTest adapter
will load later, it will clobber the other one.