Package up spec/ and upload to S3 upon releasing
Fedora has an RPM version of RR. As a part of the RPM build process, RR's tests
must be run. We no longer include tests within the gem itself so they must be
obtained from somewhere. This commit makes it so that when we release RR, the
entirety of spec/ is tar-gzipped and then uploaded to a public location (S3) so
that it can be downloaded freely.