Remove the adapter shims for Rspec and RSpec
People who use RR along with RSpec will most likely be using
`RSpec.configuration.mock_with(:rr)` instead of including an RR adapter mixin
into their test case as MiniTest or Test::Unit people are used to doing, so
there is no need to shim this. In the case of
RR::Adapters::Rspec::InvocationMatcher, we will handle this in a later commit.