Use Appraisal to run test suites
Appraisal is a much better tool for managing custom Gemfiles. Besides, our
hand-rolled didn't really work 100% to load just the gems that individual test
suites needed -- it was possible, for instance, for the wrong version of
test-unit to be loaded.

Fix running test suites so that you don't have to say `bundle exec`, this just
happens automatically.

Also, refactor some of the integration tests and push the bootstrap stuff into
the integration test helper.