New way of running/defining test suites
The list of test suites which show up as available Rake tasks is now
governed by spec/suites.yml. You can modify this file to control how
those test suites are run (supply ENV or args to the spec runner). This
file is also used to generate .travis.yml (and there's a Rake task
that you use to do this now).