Fix using RSpec's RR adapter to not override our RSpec adapter
It looks like requiring 'rr' and then using `mock_with :rr` made it so
that the have_received matcher we supply was no longer available.
This is because when our adapter loads, it tells RSpec to `mock_with`
our own module, which really tells RSpec to mix in our module into
describe blocks, when they are created. Saying `mock_with :rr` replaces
our module with RSpec's own module, which does not have_received, so
this method is no longer available in describe blocks.